Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	modified:   sparsegpt.py

+++ sparsegpt.py
@@ -177,7 +177,7 @@ class SparseGPT:
         out_features = self.layer.out_features  # Get the number of output features of the old model
         N, M = 2, 4
         s = admm_solve(W, N, M)
-        self.layer.weight.data = s
+        self.layer.weight.data = s.to(dtype=self.layer.weight.data.dtype)
         if DEBUG:
             print('error for admm:')
             print(torch.sum((self.layer(self.inp1) - self.out1) ** 2))
